Updates
Once you press the button to send a command for the first time, the permissions will be requested and once granted, the command will be sent without having to press the button twice.

Realization of the "checkPermission" method of the SMSController in the mainActivity to improve the clearness of the code.